react-text-to-speech/realTimeTest/requirements.txt
# === Core RAG stack ===
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
openai>=1.40.0
chromadb>=0.5.0
pypdf>=4.2.0

# === FastAPI server ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# === Optional but handy ===
# tiktoken is useful for token counting/chunking diagnostics
tiktoken>=0.7.0
# python-dotenv helps manage OPENAI_API_KEY in a .env file
python-dotenv>=1.0.1
